Closes #2
Because `typecheck` returns `ty option`, the `Option.map` call in `eval` was
producing a value of type `ty option option`. This means that `multistep` would
still be called even when `typecheck` produced a `None` result.

Unfortunately, the tests did not catch this behavior because, for tests that
expected to not type-check, `eval` was never called.

This commit updates the definition of `eval` to fix the unintended behavior, and
also updates the `Tests.run_test` function to check that `eval` produces `None`
when the input is expected not to type-check.